Refactor: Move code into ESQueryProfilerMixin
